pyproject.toml
[tool.poetry]
name = "django-channels-prometheus"
version = "1.0.0"
description = "Very simple package that enables a Prometheus endpoint, to check how many websocket connections are currently open."
license = "MIT"
authors = ["Frederich Pedersen <frodo@hobbits.dk>"]
repository = "https://github.com/Frodothedwarf/django-channels-prometheus"
keywords = ["prometheus","channels","webhook","webhooks","django"]
classifiers   = [
    'Development Status :: 3 - Alpha',
    'Framework :: Django',
    'Framework :: Django :: 4.0',
    'Framework :: Django :: 4.1',
    'Framework :: Django :: 4.2',
    'Framework :: Django :: 5.0',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
]
readme = "README.md"
packages      = [
    { include = "channels_prometheus" },
]

[tool.poetry.dependencies]
python = "^3.8"
channels = "*"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
tox = "^4.15.1"
pytest = "*"
pytest-django = "^4.8.0"
pytest-asyncio = "*"
daphne = "*"

[tool.poetry.extras]
test = ["pytest"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
asyncio_default_fixture_loop_scope = "class"